refactor 클래스 분리
 - User의 Input을 받는 클래스, 입력받은 값이 옳은 지 확인하는 클래스로 분리
