+++ src/main/java/calculator/Calculator.java
@@ -2,13 +2,12 @@ package calculator;
 
 import java.util.LinkedList;
 import java.util.Queue;
-import java.util.Scanner;
 
 public class Calculator {
     private Operator op = new Operator();
+    private UserInput input = new UserInput();
     private Queue<String> operatorQueue = new LinkedList<>();
     private Queue<Double> numberQueue = new LinkedList<>();
-    private String equationDelimiter = " ";
 
     public static void main(String[] args) {
         Calculator calculator = new Calculator();
@@ -16,7 +15,7 @@ public class Calculator {
     }
 
     private void run() {
-        String[] equations = getInputEquation();
+        String[] equations = input.getInputEquation();
         System.out.println(calculateEquation(equations));
     }
 
@@ -43,57 +42,5 @@ public class Calculator {
         }
 
         return result;
-    }
-
-    private String[] getInputEquation() {
-        Scanner scanner = new Scanner(System.in);
-        String[] equations;
-
-        while(true) {
-            try {
-                String equation = scanner.nextLine();
-                equations = equation.split(equationDelimiter);
-                checkEquationElementValue(equations);
-                break;
-            } catch(NumberFormatException e) {
-                System.out.println("다시 입력해주세요");
-            }
-        }
-
-        return equations;
-    }
-
-    private void checkEquationElementValue(String[] equations) {
-        checkOperatorValue(equations);
-        checkNumberValue(equations);
-    }
-
-    private void checkOperatorValue(String[] equations) {
-        for(int i=1; i<equations.length; i += 2) {
-            if(isWrongOperator(equations, i)) { throw new NumberFormatException(); }
-            if(isInvalidEquation(equations)) { throw new NumberFormatException(); }
-            if(isDividedByZero(equations, i)) { throw new NumberFormatException(); }
-        }
-    }
-
-    public boolean isWrongOperator(String[] equations, int index) {
-        return !equations[index].equals("+") && !equations[index].equals("-") && !equations[index].equals("*") && !equations[index].equals("/");
-    }
-
-    public boolean isInvalidEquation(String[] equations) {
-        return equations.length % 2 == 0;
-    }
-
-    public boolean isDividedByZero(String[] equations, int index) {
-        return equations[index].equals("/") && Double.parseDouble(equations[index+1]) == 0.0;
-    }
-
-    public void checkNumberValue(String[] equations) {
-        for(int i=0; i<equations.length; i += 2) {
-            if(Double.parseDouble(equations[i]) > (double) Integer.MAX_VALUE
-                    || Double.parseDouble(equations[i]) < -(double) Integer.MAX_VALUE) {
-                throw new NumberFormatException();
-            }
-        }
     }
 }
